Crate amethyst_error

Contains the Error type and company as used by Amethyst.

Note: This type is not intended to be used outside of Amethyst. If you are integrating a crate from amethyst to use this, it is recommended that you treat this type as an opaque std::error::Error.

Macros

format_err
Constructs an Error using the standard string interpolation syntax.

Structs

Backtrace
Representation of an owned and self-contained backtrace.

Causes
An iterator over all the causes for this error.

Error
The error type used by Amethyst.

Traits

ResultExt
Extra convenience functions for results based on core errors.

Functions

err_msg
Constructs an Error from a string.